Added new Delete Revision button
New button is active when registration is in revision state.
The new confirmation message ask to confirme the deletion of the revision (not the registration)
Removed confirmation message when clicking Revise button because there is a dialog box asking for the revision date and there is a cancel button on it.