Corrected several problems with date generation
Dates with attendances are no longer deleted in sessions and shows management.
After the dates are generated, a new message now give a summary of what was done. (deleted, inserted, etc...)
If a course or a number have no dates generated, a message is now displayed instead of a PHP error