Allow nbofdaysprior to be -1 and allow registration in the past
Allow nbofdaysprior to be -1 for a test session to make all test periods of that session opened to registration, even if period is in the past. This is to allow coaches to retroactively register the skaters' tests in the system. Test director still has to approve and grade every test, but the coaches can now help with the retro registration.